fix textures sinking

+++ includes/render.h
@@ -6,7 +6,7 @@
 /*   By: glaguyon <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/20 14:40:48 by glaguyon          #+#    #+#             */
-/*   Updated: 2024/06/27 16:31:33 by glaguyon         ###   ########.fr       */
+/*   Updated: 2024/06/27 18:24:19 by glag             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 # include "map.h"
 
 int		draw(void *data_);
-void	drawv(t_data *data, t_img img, unsigned int x, unsigned int offset, unsigned int hei);
+void	drawv(t_data *data, t_img img, unsigned int x, unsigned int offset, double hei);
 void	draw_minimap(t_data *data);
 void	move(t_data *data, float delta, uint64_t keys);
